Renamed test_helpers to helpers Added unittests for helpers
